Simplify test execution expandable and auto expand relevant sections
