+++ frontend/lib/ui/artefact_page/test_execution_expandable/test_execution_rerun_button.dart
@@ -1,0 +1,72 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:go_router/go_router.dart';
+
+import '../../../models/test_execution.dart';
+import '../../../providers/artefact_builds.dart';
+import '../../../routing.dart';
+
+class RerunButton extends ConsumerWidget {
+  const RerunButton({super.key, required this.testExecution});
+
+  final TestExecution testExecution;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final artefactId =
+        AppRoutes.artefactIdFromUri(AppRoutes.uriFromContext(context));
+
+    final handlePress = testExecution.isRerunRequested
+        ? null
+        : () => showDialog(
+              context: context,
+              builder: (_) => _RerunConfirmationDialog(
+                artefactId: artefactId,
+                testExecutionId: testExecution.id,
+              ),
+            );
+
+    return Tooltip(
+      message: testExecution.isRerunRequested ? 'Already requested' : '',
+      child: TextButton(
+        onPressed: handlePress,
+        child: const Text('rerun'),
+      ),
+    );
+  }
+}
+
+class _RerunConfirmationDialog extends ConsumerWidget {
+  const _RerunConfirmationDialog({
+    required this.artefactId,
+    required this.testExecutionId,
+  });
+
+  final int artefactId;
+  final int testExecutionId;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return AlertDialog(
+      title: const Text(
+        'Are you sure you want to rerun this environment?',
+      ),
+      actions: [
+        TextButton(
+          autofocus: true,
+          onPressed: () {
+            ref
+                .read(artefactBuildsProvider(artefactId).notifier)
+                .rerunTestExecutions({testExecutionId});
+            context.pop();
+          },
+          child: const Text('yes'),
+        ),
+        TextButton(
+          onPressed: () => context.pop(),
+          child: const Text('no'),
+        ),
+      ],
+    );
+  }
+}
